+++ uart8250/src/uart.rs
@@ -8,6 +8,8 @@ use nb;
 #[cfg(feature = "fmt")]
 use core::fmt;
 
+use crate::registers::Registers;
+
 #[derive(Debug, Clone, PartialEq)]
 pub enum ChipFifoInfo {
     NoFifo,
@@ -38,14 +40,16 @@ pub enum Parity {
 /// # MMIO version of uart 8250
 ///
 /// **Noticed** This is only tested on the NS16550 compatible UART used in QEMU 5.0 virt machine of RISC-V
-pub struct MmioUart8250 {
-    base_address: usize,
+pub struct MmioUart8250<'a> {
+    reg: &'a mut Registers,
 }
 
-impl MmioUart8250 {
+impl<'a> MmioUart8250<'a> {
     /// New a uart
     pub fn new(base_address: usize) -> Self {
-        Self { base_address }
+        Self {
+            reg: cast!(base_address),
+        }
     }
 
     /// A basic way to init the uart with interrupt enable
@@ -65,12 +69,16 @@ impl MmioUart8250 {
         // No modem control
         self.write_mcr(0);
         // Enable received_data_available_interrupt
-        self.enable_received_data_available_interrupt();
+        // Enable transmitter_holding_register_empty_interrupt
+        self.write_ier(3);
+        // same to following 2 lines
+        // self.enable_received_data_available_interrupt();
+        // self.enable_transmitter_holding_register_empty_interrupt();
     }
 
     /// Set a new base_address
     pub fn set_base_address(&mut self, base_address: usize) {
-        self.base_address = base_address;
+        self.reg = cast!(base_address);
     }
 
     /// Read a byte from uart
@@ -103,14 +111,14 @@ impl MmioUart8250 {
     /// >
     /// > If the receive buffer is occupied or the FIFO is full, the incoming data is discarded and the Receiver Line Status interrupt is written to the IIR register. The Overrun Error bit is also set in the Line Status Register.
     pub fn write_thr(&self, value: u8) {
-        unsafe { (*cast!(self.base_address)).rw[0].write(value) }
+        unsafe { self.reg.rw[0].write(value) }
     }
 
     /// read RBR (offset + 0)
     ///
     /// Read Receiver Buffer to get data
     pub fn read_rbr(&self) -> u8 {
-        unsafe { (*cast!(self.base_address)).rw[0].read() }
+        self.reg.rw[0].read()
     }
 
     /// read DLL (offset + 0)
@@ -143,39 +151,37 @@ impl MmioUart8250 {
     /// | 57600     | 2                    | $00                     | $02                    |
     /// | 115200    | 1                    | $00                     | $01                    |
     pub fn read_dll(&self) -> u8 {
-        unsafe { (*cast!(self.base_address)).rw[0].read() }
+        self.reg.rw[0].read()
     }
 
     /// write DLL (offset + 0)
     ///
     /// set divisor latch low byte in the register
     pub fn write_dll(&self, value: u8) {
-        unsafe { (*cast!(self.base_address)).rw[0].write(value) }
+        unsafe { self.reg.rw[0].write(value) }
     }
 
     /// read DLH (offset + 1)
     ///
     /// get divisor latch high byte in the register
     pub fn read_dlh(&self) -> u8 {
-        unsafe { (*cast!(self.base_address)).rw[1].read() }
+        self.reg.rw[1].read()
     }
 
     /// write DLH (offset + 1)
     ///
     /// set divisor latch high byte in the register
     pub fn write_dlh(&self, value: u8) {
-        unsafe { (*cast!(self.base_address)).rw[1].write(value) }
+        unsafe { self.reg.rw[1].write(value) }
     }
 
     /// Set divisor latch according to clock and baud_rate, then set DLAB to false
     pub fn set_divisor(&self, clock: usize, baud_rate: usize) {
-        if !self.is_divisor_latch_accessible() {
-            self.toggle_divisor_latch_accessible();
-        }
+        self.enable_divisor_latch_accessible();
         let divisor = clock / (16 * baud_rate);
         self.write_dll(divisor as u8);
         self.write_dlh((divisor >> 8) as u8);
-        self.toggle_divisor_latch_accessible();
+        self.disable_divisor_latch_accessible();
     }
 
     /// Read IER (offset + 1)
@@ -199,134 +205,134 @@ impl MmioUart8250 {
     /// > | 1   | Enable Transmitter Holding Register Empty Interrupt |
     /// > | 0   | Enable Received Data Available Interrupt            |
     pub fn read_ier(&self) -> u8 {
-        unsafe { (*cast!(self.base_address)).rw[1].read() }
+        self.reg.rw[1].read()
     }
 
     /// Write IER (offset + 1)
     ///
     /// Write Interrupt Enable Register to turn on/off interrupts
     pub fn write_ier(&self, value: u8) {
-        unsafe { (*cast!(self.base_address)).rw[1].write(value) }
+        unsafe { self.reg.rw[1].write(value) }
     }
 
     /// get whether low power mode (16750) is enabled (IER\[5\])
     pub fn is_low_power_mode_enabled(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).rw[1].read() & 0b0010_0000 != 0 }
+        self.reg.rw[1].read() & 0b0010_0000 != 0
     }
 
     /// toggle low power mode (16750) (IER\[5\])
     pub fn toggle_low_power_mode(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v ^ 0b0010_0000) }
+        unsafe { self.reg.rw[1].modify(|v| v ^ 0b0010_0000) }
     }
 
     /// enable low power mode (16750) (IER\[5\])
     pub fn enable_low_power_mode(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v | 0b0010_0000) }
+        unsafe { self.reg.rw[1].modify(|v| v | 0b0010_0000) }
     }
 
     /// disable low power mode (16750) (IER\[5\])
     pub fn disable_low_power_mode(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v & !0b0010_0000) }
+        unsafe { self.reg.rw[1].modify(|v| v & !0b0010_0000) }
     }
 
     /// get whether sleep mode (16750) is enabled (IER\[4\])
     pub fn is_sleep_mode_enabled(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).rw[1].read() & 0b0001_0000 != 0 }
+        self.reg.rw[1].read() & 0b0001_0000 != 0
     }
 
     /// toggle sleep mode (16750) (IER\[4\])
     pub fn toggle_sleep_mode(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v ^ 0b0001_0000) }
+        unsafe { self.reg.rw[1].modify(|v| v ^ 0b0001_0000) }
     }
 
     /// enable sleep mode (16750) (IER\[4\])
     pub fn enable_sleep_mode(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v | 0b0001_0000) }
+        unsafe { self.reg.rw[1].modify(|v| v | 0b0001_0000) }
     }
 
     /// disable sleep mode (16750) (IER\[4\])
     pub fn disable_sleep_mode(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v & !0b0001_0000) }
+        unsafe { self.reg.rw[1].modify(|v| v & !0b0001_0000) }
     }
 
     /// get whether modem status interrupt is enabled (IER\[3\])
     pub fn is_modem_status_interrupt_enabled(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).rw[1].read() & 0b0000_1000 != 0 }
+        self.reg.rw[1].read() & 0b0000_1000 != 0
     }
 
     /// toggle modem status interrupt (IER\[3\])
     pub fn toggle_modem_status_interrupt(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v ^ 0b0000_1000) }
+        unsafe { self.reg.rw[1].modify(|v| v ^ 0b0000_1000) }
     }
 
     /// enable modem status interrupt (IER\[3\])
     pub fn enable_modem_status_interrupt(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v | 0b0000_1000) }
+        unsafe { self.reg.rw[1].modify(|v| v | 0b0000_1000) }
     }
 
     /// disable modem status interrupt (IER\[3\])
     pub fn disable_modem_status_interrupt(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v & !0b0000_1000) }
+        unsafe { self.reg.rw[1].modify(|v| v & !0b0000_1000) }
     }
 
     /// get whether receiver line status interrupt is enabled (IER\[2\])
     pub fn is_receiver_line_status_interrupt_enabled(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).rw[1].read() & 0b0000_0100 != 0 }
+        self.reg.rw[1].read() & 0b0000_0100 != 0
     }
 
     /// toggle receiver line status interrupt (IER\[2\])
     pub fn toggle_receiver_line_status_interrupt(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v ^ 0b0000_0100) }
+        unsafe { self.reg.rw[1].modify(|v| v ^ 0b0000_0100) }
     }
 
     /// enable receiver line status interrupt (IER\[2\])
     pub fn enable_receiver_line_status_interrupt(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v | 0b0000_0100) }
+        unsafe { self.reg.rw[1].modify(|v| v | 0b0000_0100) }
     }
 
     /// disable receiver line status interrupt (IER\[2\])
     pub fn disable_receiver_line_status_interrupt(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v & !0b0000_0100) }
+        unsafe { self.reg.rw[1].modify(|v| v & !0b0000_0100) }
     }
 
     /// get whether transmitter holding register empty interrupt is enabled (IER\[1\])
     pub fn is_transmitter_holding_register_empty_interrupt_enabled(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).rw[1].read() & 0b0000_0010 != 0 }
+        self.reg.rw[1].read() & 0b0000_0010 != 0
     }
 
     /// toggle transmitter holding register empty interrupt (IER\[1\])
     pub fn toggle_transmitter_holding_register_empty_interrupt(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v ^ 0b0000_0010) }
+        unsafe { self.reg.rw[1].modify(|v| v ^ 0b0000_0010) }
     }
 
     /// enable transmitter holding register empty interrupt (IER\[1\])
     pub fn enable_transmitter_holding_register_empty_interrupt(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v | 0b0000_0010) }
+        unsafe { self.reg.rw[1].modify(|v| v | 0b0000_0010) }
     }
 
     /// disable transmitter holding register empty interrupt (IER\[1\])
     pub fn disable_transmitter_holding_register_empty_interrupt(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v & !0b0000_0010) }
+        unsafe { self.reg.rw[1].modify(|v| v & !0b0000_0010) }
     }
 
     /// get whether received data available is enabled (IER\[0\])
     pub fn is_received_data_available_interrupt_enabled(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).rw[1].read() & 0b0000_0001 != 0 }
+        self.reg.rw[1].read() & 0b0000_0001 != 0
     }
 
     /// toggle received data available (IER\[0\])
     pub fn toggle_received_data_available_interrupt(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v ^ 0b0000_0001) }
+        unsafe { self.reg.rw[1].modify(|v| v ^ 0b0000_0001) }
     }
 
     /// enable received data available (IER\[0\])
     pub fn enable_received_data_available_interrupt(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v | 0b0000_0001) }
+        unsafe { self.reg.rw[1].modify(|v| v | 0b0000_0001) }
     }
 
     /// disable received data available (IER\[0\])
     pub fn disable_received_data_available_interrupt(&self) {
-        unsafe { (*cast!(self.base_address)).rw[1].modify(|v| v & !0b0000_0001) }
+        unsafe { self.reg.rw[1].modify(|v| v & !0b0000_0001) }
     }
 
     /// Read IIR (offset + 2)
@@ -362,12 +368,12 @@ impl MmioUart8250 {
     /// > |            | 1                                 | 1     | 1                                 | Reserved                                     | N/A                                                                                       |
     /// > | 0          | Interrupt Pending Flag            |       |                                   |                                              |                                                                                           |
     pub fn read_iir(&self) -> u8 {
-        unsafe { (*cast!(self.base_address)).rw[2].read() }
+        self.reg.rw[2].read()
     }
 
     /// Read IIR\[7:6\] to get FIFO status
     pub fn read_fifo_status(&self) -> ChipFifoInfo {
-        match unsafe { (*cast!(self.base_address)).rw[2].read() & 0b1100_0000 } {
+        match self.reg.rw[2].read() & 0b1100_0000 {
             0 => ChipFifoInfo::NoFifo,
             0b0100_0000 => ChipFifoInfo::Reserved,
             0b1000_0000 => ChipFifoInfo::EnabledNoFunction,
@@ -378,12 +384,12 @@ impl MmioUart8250 {
 
     /// get whether 64 Byte fifo (16750 only) is enabled (IIR\[5\])
     pub fn is_64byte_fifo_enabled(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).rw[2].read() & 0b0010_0000 != 0 }
+        self.reg.rw[2].read() & 0b0010_0000 != 0
     }
 
     /// Read IIR\[3:1\] to get interrupt type
     pub fn read_interrupt_type(&self) -> InterruptType {
-        match unsafe { (*cast!(self.base_address)).rw[2].read() & 0b0000_1110 } {
+        match self.reg.rw[2].read() & 0b0000_1110 {
             0 => InterruptType::ModemStatus,
             0b0010 => InterruptType::TransmitterHoldingRegisterEmpty,
             0b0100 => InterruptType::ReceivedDataAvailable,
@@ -396,7 +402,7 @@ impl MmioUart8250 {
 
     /// get whether interrupt is pending (IIR\[0\])
     pub fn is_interrupt_pending(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).rw[2].read() & 1 != 0 }
+        self.reg.rw[2].read() & 1 != 0
     }
 
     /// Write FCR (offset + 2) to control FIFO buffers
@@ -421,7 +427,7 @@ impl MmioUart8250 {
     /// > | 1     | Clear Receive FIFO          |       |                                   |                         |
     /// > | 0     | Enable FIFOs                |       |                                   |                         |
     pub fn write_fcr(&self, value: u8) {
-        unsafe { (*cast!(self.base_address)).rw[2].write(value) }
+        unsafe { self.reg.rw[2].write(value) }
     }
 
     /// Read LCR (offset + 3)
@@ -453,39 +459,39 @@ impl MmioUart8250 {
     /// > |          | 1                        | 0                            | 7 Bits      |               |
     /// > |          | 1                        | 1                            | 8 Bits      |               |
     pub fn read_lcr(&self) -> u8 {
-        unsafe { (*cast!(self.base_address)).rw[3].read() }
+        self.reg.rw[3].read()
     }
 
     /// Write LCR (offset + 3)
     ///
     /// Write Line Control Register to set DLAB and the serial data protocol
     pub fn write_lcr(&self, value: u8) {
-        unsafe { (*cast!(self.base_address)).rw[3].write(value) }
+        unsafe { self.reg.rw[3].write(value) }
     }
 
     /// get whether DLAB is enabled
     pub fn is_divisor_latch_accessible(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).rw[3].read() & 0b1000_0000 != 0 }
+        self.reg.rw[3].read() & 0b1000_0000 != 0
     }
 
     /// toggle DLAB
     pub fn toggle_divisor_latch_accessible(&self) {
-        unsafe { (*cast!(self.base_address)).rw[3].modify(|v| v ^ 0b1000_0000) }
+        unsafe { self.reg.rw[3].modify(|v| v ^ 0b1000_0000) }
     }
 
     /// enable DLAB
     pub fn enable_divisor_latch_accessible(&self) {
-        unsafe { (*cast!(self.base_address)).rw[3].modify(|v| v | 0b1000_0000) }
+        unsafe { self.reg.rw[3].modify(|v| v | 0b1000_0000) }
     }
 
     /// disable DLAB
     pub fn disable_divisor_latch_accessible(&self) {
-        unsafe { (*cast!(self.base_address)).rw[3].modify(|v| v & !0b1000_0000) }
+        unsafe { self.reg.rw[3].modify(|v| v & !0b1000_0000) }
     }
 
     /// get parity of used data protocol
     pub fn get_parity(&self) -> Parity {
-        match unsafe { (*cast!(self.base_address)).rw[3].read() & 0b0011_1000 } {
+        match self.reg.rw[3].read() & 0b0011_1000 {
             0b0000_0000 => Parity::No,
             0b0000_1000 => Parity::Odd,
             0b0001_1000 => Parity::Even,
@@ -498,21 +504,11 @@ impl MmioUart8250 {
     /// set parity
     pub fn set_parity(&self, parity: Parity) {
         match parity {
-            Parity::No => unsafe {
-                (*cast!(self.base_address)).rw[3].modify(|v| (v & 0b1100_0111))
-            },
-            Parity::Odd => unsafe {
-                (*cast!(self.base_address)).rw[3].modify(|v| (v & 0b1100_0111) | 0b0000_1000)
-            },
-            Parity::Even => unsafe {
-                (*cast!(self.base_address)).rw[3].modify(|v| (v & 0b1100_0111) | 0b0001_1000)
-            },
-            Parity::Mark => unsafe {
-                (*cast!(self.base_address)).rw[3].modify(|v| (v & 0b1100_0111) | 0b0010_1000)
-            },
-            Parity::Space => unsafe {
-                (*cast!(self.base_address)).rw[3].modify(|v| v | 0b0011_1000)
-            },
+            Parity::No => unsafe { self.reg.rw[3].modify(|v| (v & 0b1100_0111)) },
+            Parity::Odd => unsafe { self.reg.rw[3].modify(|v| (v & 0b1100_0111) | 0b0000_1000) },
+            Parity::Even => unsafe { self.reg.rw[3].modify(|v| (v & 0b1100_0111) | 0b0001_1000) },
+            Parity::Mark => unsafe { self.reg.rw[3].modify(|v| (v & 0b1100_0111) | 0b0010_1000) },
+            Parity::Space => unsafe { self.reg.rw[3].modify(|v| v | 0b0011_1000) },
         }
     }
 
@@ -520,27 +516,27 @@ impl MmioUart8250 {
     ///
     /// Simply return a u8 to indicate 1 or 1.5/2 bits
     pub fn get_stop_bit(&self) -> u8 {
-        ((unsafe { (*cast!(self.base_address)).rw[3].read() & 0b100 }) >> 2) + 1
+        ((self.reg.rw[3].read() & 0b100) >> 2) + 1
     }
 
     /// set stop bit, only 1 and 2 can be used as `stop_bit`
     pub fn set_stop_bit(&self, stop_bit: u8) {
         match stop_bit {
-            1 => unsafe { (*cast!(self.base_address)).rw[3].modify(|v| v & 0b1111_1011) },
-            2 => unsafe { (*cast!(self.base_address)).rw[3].modify(|v| v | 0b0000_0100) },
+            1 => unsafe { self.reg.rw[3].modify(|v| v & 0b1111_1011) },
+            2 => unsafe { self.reg.rw[3].modify(|v| v | 0b0000_0100) },
             _ => panic!("Invalid stop bit"),
         }
     }
 
     /// get word length of used data protocol
     pub fn get_word_length(&self) -> u8 {
-        (unsafe { (*cast!(self.base_address)).rw[3].read() & 0b11 }) + 5
+        (self.reg.rw[3].read() & 0b11) + 5
     }
 
     /// set word length, only 5..=8 can be used as `length`
     pub fn set_word_length(&self, length: u8) {
         if (5..=8).contains(&length) {
-            unsafe { (*cast!(self.base_address)).rw[3].modify(|v| v | (length - 5)) }
+            unsafe { self.reg.rw[3].modify(|v| v | (length - 5)) }
         } else {
             panic!("Invalid word length")
         }
@@ -565,14 +561,14 @@ impl MmioUart8250 {
     /// > | 1   | Request To Send                  |
     /// > | 0   | Data Terminal Ready              |
     pub fn read_mcr(&self) -> u8 {
-        unsafe { (*cast!(self.base_address)).rw[4].read() }
+        self.reg.rw[4].read()
     }
 
     /// Write MCR (offset + 4)
     ///
     /// Write Modem Control Register to control flow
     pub fn write_mcr(&self, value: u8) {
-        unsafe { (*cast!(self.base_address)).rw[4].write(value) }
+        unsafe { self.reg.rw[4].write(value) }
     }
 
     /// Read LSR (offset + 5)
@@ -592,42 +588,42 @@ impl MmioUart8250 {
     /// > | 1   | Overrun Error                      |
     /// > | 0   | Data Ready                         |
     pub fn read_lsr(&self) -> u8 {
-        unsafe { (*cast!(self.base_address)).ro[0].read() }
+        self.reg.ro[0].read()
     }
 
     /// get whether there is an error in received FIFO
     pub fn is_received_fifo_error(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[0].read() & 0b1000_0000 != 0 }
+        self.reg.ro[0].read() & 0b1000_0000 != 0
     }
 
     /// get whether data holding registers are empty
     pub fn is_data_holding_registers_empty(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[0].read() & 0b0100_0000 != 0 }
+        self.reg.ro[0].read() & 0b0100_0000 != 0
     }
 
     /// get whether transmitter holding register is empty
     pub fn is_transmitter_holding_register_empty(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[0].read() & 0b0010_0000 != 0 }
+        self.reg.ro[0].read() & 0b0010_0000 != 0
     }
 
     pub fn is_break_interrupt(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[0].read() & 0b0001_0000 != 0 }
+        self.reg.ro[0].read() & 0b0001_0000 != 0
     }
 
     pub fn is_framing_error(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[0].read() & 0b0000_1000 != 0 }
+        self.reg.ro[0].read() & 0b0000_1000 != 0
     }
 
     pub fn is_parity_error(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[0].read() & 0b0000_0100 != 0 }
+        self.reg.ro[0].read() & 0b0000_0100 != 0
     }
 
     pub fn is_overrun_error(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[0].read() & 0b0000_0010 != 0 }
+        self.reg.ro[0].read() & 0b0000_0010 != 0
     }
 
     pub fn is_data_ready(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[0].read() & 0b0000_0001 != 0 }
+        self.reg.ro[0].read() & 0b0000_0001 != 0
     }
 
     /// Read MSR (offset + 6)
@@ -647,47 +643,47 @@ impl MmioUart8250 {
     /// > | 1   | Delta Data Set Ready         |
     /// > | 0   | Delta Clear To Send          |
     pub fn read_msr(&self) -> u8 {
-        unsafe { (*cast!(self.base_address)).ro[1].read() }
+        self.reg.ro[1].read()
     }
 
     pub fn is_carrier_detect(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[1].read() & 0b1000_0000 != 0 }
+        self.reg.ro[1].read() & 0b1000_0000 != 0
     }
 
     pub fn is_ring_indicator(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[1].read() & 0b0100_0000 != 0 }
+        self.reg.ro[1].read() & 0b0100_0000 != 0
     }
 
     pub fn is_data_set_ready(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[1].read() & 0b0010_0000 != 0 }
+        self.reg.ro[1].read() & 0b0010_0000 != 0
     }
 
     pub fn is_clear_to_send(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[1].read() & 0b0001_0000 != 0 }
+        self.reg.ro[1].read() & 0b0001_0000 != 0
     }
 
     pub fn is_delta_data_carrier_detect(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[1].read() & 0b0000_1000 != 0 }
+        self.reg.ro[1].read() & 0b0000_1000 != 0
     }
 
     pub fn is_trailing_edge_ring_indicator(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[1].read() & 0b0000_0100 != 0 }
+        self.reg.ro[1].read() & 0b0000_0100 != 0
     }
 
     pub fn is_delta_data_set_ready(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[1].read() & 0b0000_0010 != 0 }
+        self.reg.ro[1].read() & 0b0000_0010 != 0
     }
 
     pub fn is_delta_clear_to_send(&self) -> bool {
-        unsafe { (*cast!(self.base_address)).ro[1].read() & 0b0000_0001 != 0 }
+        self.reg.ro[1].read() & 0b0000_0001 != 0
     }
 
     pub fn read_sr(&self) -> u8 {
-        unsafe { (*cast!(self.base_address)).scratch.read() }
+        self.reg.scratch.read()
     }
 
     pub fn write_sr(&self, value: u8) {
-        unsafe { (*cast!(self.base_address)).scratch.write(value) }
+        unsafe { self.reg.scratch.write(value) }
     }
 }
 
@@ -695,7 +691,7 @@ impl MmioUart8250 {
 ///
 /// This is a very simple implementation, based on [rustsbi/rustsbi-qemu](https://github.com/rustsbi/rustsbi-qemu/blob/main/rustsbi-qemu/src/ns16550a.rs#L36-L52)
 #[cfg(feature = "embedded")]
-impl serial::Read<u8> for MmioUart8250 {
+impl<'a> serial::Read<u8> for MmioUart8250<'a> {
     type Error = Infallible;
 
     fn try_read(&mut self) -> nb::Result<u8, Self::Error> {
@@ -711,7 +707,7 @@ impl serial::Read<u8> for MmioUart8250 {
 ///
 /// This is a very simple implementation, based on [rustsbi/rustsbi-qemu](https://github.com/rustsbi/rustsbi-qemu/blob/main/rustsbi-qemu/src/ns16550a.rs#L54-L75)
 #[cfg(feature = "embedded")]
-impl serial::Write<u8> for MmioUart8250 {
+impl<'a> serial::Write<u8> for MmioUart8250<'a> {
     type Error = Infallible;
 
     fn try_write(&mut self, word: u8) -> nb::Result<(), Self::Error> {
@@ -732,7 +728,7 @@ impl serial::Write<u8> for MmioUart8250 {
 ///
 /// A simple implementation, may be changed in the future
 #[cfg(feature = "fmt")]
-impl fmt::Write for MmioUart8250 {
+impl<'a> fmt::Write for MmioUart8250<'a> {
     fn write_str(&mut self, s: &str) -> fmt::Result {
         for c in s.as_bytes() {
             self.write_thr(*c);